refactor: 폴더명 book -> gig 으로 변경
