+++ src/pages/gig/components/content/Content.styled.ts
