+++ src/pages/gig/components/content/Content.tsx
