+++ src/pages/gig/components/iconText/IconText.tsx
