+++ src/pages/gig/components/peopleCard/PeopleCard.styled.ts
