+++ src/pages/gig/components/peopleCard/PeopleCard.tsx
