+++ src/pages/gig/components/showInfo/ShowInfo.styled.ts
