+++ src/pages/gig/components/showInfo/ShowInfo.tsx
