+++ src/pages/gig/components/tabBar/TabBar.styled.ts
